+++ apps/api/src/auth/controllers/api-key/api-key.controller.ts
@@ -1,7 +1,13 @@
 import assert from "http-assert";
 import { singleton } from "tsyringe";
 
-import { CreateApiKeyRequest, ListApiKeysResponse, SingleApiKeyResponse, UpdateApiKeyRequest } from "@src/auth/http-schemas/api-key.schema";
+import {
+  ApiKeyHiddenResponse,
+  ApiKeyVisibleResponse,
+  CreateApiKeyRequest,
+  ListApiKeysResponse,
+  UpdateApiKeyRequest
+} from "@src/auth/http-schemas/api-key.schema";
 import { ApiKeyService } from "@src/auth/services/api-key/api-key.service";
 import { Protected } from "@src/auth/services/auth.service";
 
@@ -16,20 +22,20 @@ export class ApiKeyController {
   }
 
   @Protected([{ action: "read", subject: "ApiKey" }])
-  async findById(id: string): Promise<SingleApiKeyResponse> {
+  async findById(id: string): Promise<ApiKeyHiddenResponse> {
     const apiKey = await this.apiKeyService.findById(id);
     assert(apiKey, 404, "API key not found");
     return { data: apiKey };
   }
 
   @Protected([{ action: "create", subject: "ApiKey" }])
-  async create(input: CreateApiKeyRequest["data"]): Promise<SingleApiKeyResponse> {
+  async create(input: CreateApiKeyRequest["data"]): Promise<ApiKeyVisibleResponse> {
     const apiKey = await this.apiKeyService.create(input);
     return { data: apiKey };
   }
 
   @Protected([{ action: "update", subject: "ApiKey" }])
-  async update(id: string, input: UpdateApiKeyRequest["data"]): Promise<SingleApiKeyResponse> {
+  async update(id: string, input: UpdateApiKeyRequest["data"]): Promise<ApiKeyHiddenResponse> {
     const apiKey = await this.apiKeyService.update(id, input);
     assert(apiKey, 404, "API key not found");
     return { data: apiKey };