+++ apps/api/src/deployment/services/lease/lease.service.ts
@@ -1,9 +1,10 @@
 import { singleton } from "tsyringe";
 
-import { UserWalletOutput } from "@src/billing/repositories";
 import { ManagedSignerService, RpcMessageService } from "@src/billing/services";
+import { WalletReaderService } from "@src/billing/services/wallet-reader/wallet-reader.service";
 import { GetDeploymentResponse } from "@src/deployment/http-schemas/deployment.schema";
 import { CreateLeaseRequest } from "@src/deployment/http-schemas/lease.schema";
+import { LeaseRepository } from "@src/deployment/repositories/lease/lease.repository";
 import { ProviderService } from "@src/provider/services/provider/provider.service";
 import { DeploymentReaderService } from "../deployment-reader/deployment-reader.service";
 
@@ -13,11 +14,20 @@ export class LeaseService {
     private readonly signerService: ManagedSignerService,
     private readonly rpcMessageService: RpcMessageService,
     private readonly providerService: ProviderService,
-    private readonly deploymentReaderService: DeploymentReaderService
+    private readonly deploymentReaderService: DeploymentReaderService,
+    private readonly leaseRepository: LeaseRepository,
+    private readonly walletReaderService: WalletReaderService
   ) {}
 
-  public async createLeasesAndSendManifest(wallet: UserWalletOutput, input: CreateLeaseRequest): Promise<GetDeploymentResponse["data"]> {
-    const leaseMessages = input.leases.map(lease =>
+  public async createLeasesAndSendManifest({
+    leases,
+    manifest,
+    certificate,
+    userId
+  }: CreateLeaseRequest & { userId: string }): Promise<GetDeploymentResponse["data"]> {
+    const wallet = await this.walletReaderService.getWalletByUserId(userId);
+
+    const leaseMessages = leases.map(lease =>
       this.rpcMessageService.getCreateLeaseMsg({
         owner: wallet.address!,
         dseq: lease.dseq,
@@ -29,15 +39,18 @@ export class LeaseService {
 
     await this.signerService.executeDecodedTxByUserId(wallet.userId, leaseMessages);
 
-    for (const lease of input.leases) {
-      await this.providerService.sendManifest(lease.provider, lease.dseq, input.manifest, {
-        certPem: input.certificate.certPem,
-        keyPem: input.certificate.keyPem
+    for (const lease of leases) {
+      const commonParams = {
+        provider: lease.provider,
+        dseq: lease.dseq,
+        manifest: manifest
+      };
+      await this.providerService.sendManifest({
+        ...commonParams,
+        auth: await this.providerService.toProviderAuth(certificate || { walletId: wallet.id, provider: lease.provider })
       });
     }
 
-    return await this.deploymentReaderService.findByOwnerAndDseq(wallet.address!, input.leases[0].dseq, {
-      certificate: { certPem: input.certificate.certPem, keyPem: input.certificate.keyPem }
-    });
+    return await this.deploymentReaderService.findByWalletAndDseq(wallet, leases[0].dseq);
   }
 }